firebase_authtecation/requirements.txt
wmi
firebase_admin
python-dotenv
pyinstaller